Resolves Go module new paths and adds k8s deployment objects
