+++ Dockerfile
@@ -1,0 +1,3 @@
+FROM scratch
+ADD ./build/ca /
+CMD ["/ca"]
