+++ go.mod
@@ -1,0 +1,21 @@
+module github.com/lamassuiot/lamassu-ca
+
+go 1.13
+
+require (
+	github.com/dgrijalva/jwt-go v3.2.0+incompatible
+	github.com/go-kit/kit v0.10.0
+	github.com/go-stack/stack v1.8.0 // indirect
+	github.com/gorilla/mux v1.7.4
+	github.com/hashicorp/consul/api v1.4.0
+	github.com/hashicorp/vault v1.6.0
+	github.com/hashicorp/vault/api v1.0.5-0.20201001211907-38d91b749c77
+	github.com/hashicorp/vault/sdk v0.1.14-0.20201109203410-5e6e24692b32
+	github.com/kelseyhightower/envconfig v1.4.0
+	github.com/lib/pq v1.8.0
+	github.com/nvellon/hal v0.3.0
+	github.com/opentracing/opentracing-go v1.1.0
+	github.com/prometheus/client_golang v1.8.0
+	github.com/uber/jaeger-client-go v2.25.0+incompatible // indirect
+	github.com/uber/jaeger-lib v2.4.0+incompatible // indirect
+)
