+++ pkg/auth/testdata/keycloak.key
