+++ pkg/utils/utils.go
@@ -1,0 +1,34 @@
+package utils
+
+import (
+	"crypto/rsa"
+	"crypto/x509"
+	"encoding/pem"
+	"errors"
+	"io/ioutil"
+)
+
+const (
+	PublicKeyHeader = "-----BEGIN PUBLIC KEY-----"
+	PublicKeyFooter = "-----END PUBLIC KEY-----"
+)
+
+func CreateCAPool(CAPath string) (*x509.CertPool, error) {
+	caCert, err := ioutil.ReadFile(CAPath)
+	if err != nil {
+		return nil, err
+	}
+	caCertPool := x509.NewCertPool()
+	caCertPool.AppendCertsFromPEM(caCert)
+	return caCertPool, nil
+}
+
+func ParseKeycloakPublicKey(data []byte) (*rsa.PublicKey, error) {
+	pubPem, _ := pem.Decode(data)
+	parsedKey, err := x509.ParsePKIXPublicKey(pubPem.Bytes)
+	if err != nil {
+		return nil, errors.New("Unable to parse public key")
+	}
+	pubKey := parsedKey.(*rsa.PublicKey)
+	return pubKey, nil
+}
